requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2
djangorestframework==3.14.0
idna==3.4
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pytz==2023.3
requests==2.30.0
sqlparse==0.4.4
urllib3==2.0.2
